Add non-subfloor versions of planet/asteroid tiles.
We are using these for natural areas in the station, but as they are marked as subfloor, pipes/wires show above them. These non-subfloor versions may be found in the tiles dialog (F6) by searching for 'non-subfloor'. They are currently able to be removed with prying, with the thought that they are essentially trays on which the material sits. When we are further along with spriting, there may be a separate branch for creating sprites of these for when they are pried up (e.g: when you pry up sand tiles, they show up as 2x2 steel tiles because there aren't sprites for those entities).